pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "glimcrawl"
version = "0.1.0"
description = "一个强大的Google图片爬虫工具"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = ["crawler", "image", "google", "downloader"]
authors = [
    {name = "Duckweeds7", email = "root@duckweeds7.com"}
]
dependencies = [
    "loguru>=0.7.2",
    "playwright>=1.41.2",
    "aiohttp>=3.9.1",
    "aiofiles>=23.2.1",
    "Pillow>=10.0.0",
    "click>=8.0.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[project.urls]
Homepage = "https://gitee.com/duckweeds7/glimcrawl"
Repository = "https://gitee.com/duckweeds7/glimcrawl.git"
Documentation = "https://gitee.com/duckweeds7/glimcrawl#readme"

[project.scripts]
glimcrawl = "glimcrawl.cli:main"

[tool.setuptools]
packages = ["glimcrawl"]

[tool.setuptools_scm]
write_to = "glimcrawl/_version.py"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q" 